multi-pm sidebar: get composing working
